+++ Laravel/app/Http/Controllers/Manager/Branches/BranchesListing.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Manager\Branches;
 
+use App\Http\Controllers\Controller;
 use App\Models\Branch;
 
 class BranchesListing extends Controller
 {
-    public function index()
+    public function __invoke()
     {
         $branches = Branch::paginate(10);
